+++ HISTORY.md
@@ -1,3 +1,7 @@
+## 1.4.2 (03.05.2014)
+* [Add] В клаcc DLTemplate добавлен новый метод parseDocumentSource который делает замену MODX плейсхолдеров в документе
+* [Refactor] Результатирующий ответ от DocLister пропускается через DLTemplate::parseDocumentSource
+
 ## 1.4.1 (03.05.2014)
 * [Bug] Переименовывание класса paginate в DLpaginate